+++ src/types/device/EnvironmentSensor.ts
@@ -7,16 +7,26 @@ import type {
 } from './Device'
 
 export interface EnvironmentSensorAttributes
-  extends CommonDeviceAttributes,
+  extends
+    CommonDeviceAttributes,
     IdentifiableDeviceAttributes,
     JoinableDeviceAttributes,
     OtaUpdatableDeviceAttributes {
-  currentTemperature: number
-  currentRH: number
-  currentPM25: number
-  maxMeasuredPM25: number
-  minMeasuredPM25: number
-  vocIndex: number
+  currentTemperature?: number
+  currentRH?: number
+  currentPM25?: number
+  maxMeasuredPM25?: number
+  minMeasuredPM25?: number
+  vocIndex?: number
+  isOn?: boolean
+  currentCO2?: number
+  maxMeasuredCO2?: number
+  minMeasuredCO2?: number
+  discriminator?: number
+  qrCode?: string
+  setupCode?: string
+  windowOpen?: boolean
+  batteryPercentage?: number
 }
 
 export interface EnvironmentSensor extends Device {